TC#47: Renamed tests names in script
