+++ frontend/src/features/tournaments/tournamentsThunks.ts
@@ -2,7 +2,7 @@ import { createAsyncThunk } from '@reduxjs/toolkit';
 import { axiosApi } from '@/axiosApi';
 import { isAxiosError } from 'axios';
 import { GlobalError } from '@/types/userTypes';
-import { TournamentMutation, Tournaments, UpdateTournamentArg } from '@/types/tournamentTypes';
+import { TournamentMutation, Tournaments, UpdateTournamentArg } from '@/types/tournament';
 
 export const fetchTournaments = createAsyncThunk<Tournaments, string | undefined>(
   'tournaments/fetchAll',
