+++ frontend/src/types/tournament.d.ts
