+++ tests/tests.sh
@@ -24,13 +24,13 @@ echo '### Running fixtures'
 npm run seed:test
 
 echo '### Running API server in test mode'
-pm2 start "npm run start:test" --name="shop-api-test"
+pm2 start "npm run start:test" --name="API-tests"
 
 echo '### Frontend'
 cd ../frontend || exit 1
 
 echo '### Running Frontend in test mode'
-pm2 start "npm run start:test" --name="shop-frontend-test"
+pm2 start "npm run start:test" --name="Frontend-tests"
 
 while ! nc -z localhost 5183; do
   sleep 0.1
